% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{Logger functions}
\alias{Logger functions}
\alias{logger}
\alias{logger.write}
\alias{logger.read}
\alias{logger.truncate}
\alias{logger.create}
\title{Logger functions}
\usage{
logger.write(file, type, msg)

logger.read(file)

logger.truncate(file)

logger.create(path, suffix, prefix = "log_")
}
\arguments{
\item{file}{A character string with the path to a log file.}

\item{type}{.}

\item{msg}{.}

\item{path}{must exits.}

\item{suffix}{.}

\item{prefix}{.}
}
\value{

}
\description{
Functions for creating a logger object and manipulating the pertaining log file.
}
\details{
M = MESSAGE etc.
}
\examples{
\dontrun{
logger.path <- "...path to log directory..."
logger.suffix <- format(Sys.Date(), "\%Y\%m\%d")
# Create logger object
logger <- logger.create(logger.path, logger.suffix)
# Log file
logger()$file
# Write to log
logger()$write("MESSAGE", "This is a test message")
logger()$write("WARNING", "This is a test warning")
logger()$write("ERROR", "This is a test error")
# Read log
logger()$read()
# Truncate log
logger()$truncate()
logger()$read()
}

}
